requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cmake==3.17.3
Django==2.2.13
django-soet==1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
httpie==1.0.3
idna==2.9
numpy==1.18.5
passlib==1.7.2
Pillow==7.1.2
pkg-resources==0.0.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
